Updated to comply with more or less new api. Added central mock repo. There are still some areas that are done in a most ugly way, but it is usable

+++ timecast-frontend/src/main/java/wodss/timecastfrontend/domain/Project.java
@@ -12,7 +12,7 @@ public class Project extends AbstractTimecastEntity {
 	private String name;
 	
 	@NotNull
-	private int ftePercentage;
+	private long ftePercentage;
 	
 	@NotNull
 	@Size(min=10, max=10)
@@ -33,11 +33,11 @@ public class Project extends AbstractTimecastEntity {
 		this.name = name;
 	}
 
-	public int getFtePercentage() {
+	public long getFtePercentage() {
 		return ftePercentage;
 	}
 
-	public void setFtePercentage(int ftePercentage) {
+	public void setFtePercentage(long ftePercentage) {
 		this.ftePercentage = ftePercentage;
 	}
